AgroBot/requirements.txt
# AgroBot Requirements
# =====================
# Install with: pip install -r requirements.txt
# NOTE: Python 3.11 recommended (faster-whisper compatibility)

# ── Core LLM & Orchestration ──────────────────────────────────
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-groq>=0.2.0
langgraph>=0.2.0
groq>=0.9.0
python-dotenv>=1.0.0

# ── ML Models ─────────────────────────────────────────────────
xgboost>=2.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

# ── Voice: STT & TTS ──────────────────────────────────────────
edge-tts>=6.1.0
gtts>=2.5.0
pyttsx3>=2.90
sounddevice>=0.4.6
scipy>=1.11.0

# ── Image Processing ──────────────────────────────────────────
Pillow>=10.0.0
requests>=2.31.0

# ── CLI / Terminal ────────────────────────────────────────────
rich>=13.7.0
prompt_toolkit>=3.0.0

# ── Language Detection ────────────────────────────────────────
langdetect>=1.0.9

# ── Vector Store & Embeddings (for RAG) ───────────────────────
chromadb>=0.5.0
langchain-chroma>=0.1.0
sentence-transformers>=3.0.0

# ── Web Search ────────────────────────────────────────────────
duckduckgo-search>=6.0.0

# ── Explainability ────────────────────────────────────────────
shap>=0.44.0

# ── Optional: SHAP visualization ──────────────────────────────
matplotlib>=3.7.0
seaborn>=0.12.0

# ── API ───────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn>=0.29.0
gunicorn>=21.2.0
python-multipart>=0.0.9